+++ container_guts/client/diff.py
@@ -1,16 +1,16 @@
 __author__ = "Vanessa Sochat"
-__copyright__ = "Copyright 2022, Vanessa Sochat"
+__copyright__ = "Copyright 2022-2024, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
 import json
 import os
 
 import container_guts.utils as utils
+
 from ..main import ManifestGenerator
 
 
 def main(args, parser, extra, subparser):
-
     # Show args to the user
     print("         image: %s" % args.image)
     print("       outfile: %s" % args.outfile)