+++ container_guts/defaults.py
@@ -1,9 +1,10 @@
 __author__ = "Vanessa Sochat"
-__copyright__ = "Copyright 2021-2022, Vanessa Sochat"
+__copyright__ = "Copyright 2021-2024, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-import container_guts.utils as utils
 import os
+
+import container_guts.utils as utils
 
 # Default database for base image
 database = "https://github.com/singularityhub/shpc-guts"