+++ container_guts/main/container/decorator.py
@@ -1,6 +1,6 @@
 __author__ = "Vanessa Sochat"
-__copyright__ = "Copyright The ORAS Authors."
-__license__ = "Apache-2.0"
+__copyright__ = "Copyright 2021-2024, Vanessa Sochat"
+__license__ = "MPL 2.0"
 
 from functools import partial, update_wrapper