+++ docs/conf.py
@@ -13,17 +13,17 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
-import sys
 import os
+import sys
+
 from recommonmark.parser import CommonMarkParser
-
-source_parsers = {".md": CommonMarkParser}
-
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath("../"))
+
+source_parsers = {".md": CommonMarkParser}
 
 # -- General configuration ------------------------------------------------
 
@@ -60,7 +60,7 @@ master_doc = "index"
 project = "Container Guts 'Guts'"
 copyright = "2022-2023, Vanessa Sochat"
 
-from container_guts import version
+from container_guts import version  # noqa
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the